refactor: (#151) AvailableService~~ -> AvailableFeature~~
