+++ dms-domain/src/main/kotlin/team/aliens/dms/domain/school/model/AvailableFeature.kt
@@ -1,10 +1,10 @@
-package team.aliens.dms.domain.feature
+package team.aliens.dms.domain.school.model
 
 import team.aliens.dms.common.annotation.Aggregate
 import java.util.UUID
 
 @Aggregate
-data class AvailableService(
+data class AvailableFeature(
 
     val schoolId: UUID,
