+++ dms-infrastructure/src/main/kotlin/team/aliens/dms/persistence/school/mapper/AvailableFeatureMapper.kt
@@ -1,20 +1,20 @@
-package team.aliens.dms.persistence.feature.mapper
+package team.aliens.dms.persistence.school.mapper
 
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Component
-import team.aliens.dms.domain.feature.AvailableService
+import team.aliens.dms.domain.school.model.AvailableFeature
 import team.aliens.dms.persistence.GenericMapper
 import team.aliens.dms.persistence.school.repository.SchoolJpaRepository
-import team.aliens.dms.persistence.feature.entity.AvailableServiceJpaEntity
+import team.aliens.dms.persistence.school.entity.AvailableFeatureJpaEntity
 
 @Component
-class AvailableServiceMapper(
+class AvailableFeatureMapper(
     private val schoolJpaRepository: SchoolJpaRepository
-) : GenericMapper<AvailableService, AvailableServiceJpaEntity> {
+) : GenericMapper<AvailableFeature, AvailableFeatureJpaEntity> {
 
-    override fun toDomain(entity: AvailableServiceJpaEntity?): AvailableService? {
+    override fun toDomain(entity: AvailableFeatureJpaEntity?): AvailableFeature? {
         return entity?.let {
-            AvailableService(
+            AvailableFeature(
                 schoolId = entity.schoolId,
                 mealService = entity.mealService,
                 noticeService = entity.noticeService,
@@ -23,10 +23,10 @@ class AvailableServiceMapper(
         }
     }
 
-    override fun toEntity(domain: AvailableService): AvailableServiceJpaEntity {
+    override fun toEntity(domain: AvailableFeature): AvailableFeatureJpaEntity {
         val school = schoolJpaRepository.findByIdOrNull(domain.schoolId)
 
-        return AvailableServiceJpaEntity(
+        return AvailableFeatureJpaEntity(
             schoolId = domain.schoolId,
             school = school,
             mealService = domain.mealService,
