+++ dms-infrastructure/src/main/kotlin/team/aliens/dms/persistence/school/repository/AvailableFeatureJpaRepository.kt
@@ -1,0 +1,10 @@
+package team.aliens.dms.persistence.school.repository
+
+import org.springframework.data.repository.CrudRepository
+import org.springframework.stereotype.Repository
+import team.aliens.dms.persistence.school.entity.AvailableFeatureJpaEntity
+import java.util.UUID
+
+@Repository
+interface AvailableFeatureJpaRepository : CrudRepository<AvailableFeatureJpaEntity, UUID> {
+}
